Implement branch-and-bound
This adds an experimental branch and bound solver to the addon. It is
designed to optimize the current worst case for the dynamic programming
solver, namely when two caps are applied. This solver runs in less time
(though not instant) and also uses much less memory - from over 2GB peak
allocation with the current solver to ~85MB with branch and bound.

We traverse the items in a depth-first manner and prune subtrees using
two techniques. Both of these techniques rely on a precomputed suffix
list, which calculates the cumulative maximum and minimum cap
contributions and maximum achievable non-cap score of the last 1-15
items. When assessing whether to continue traversing a branch, we pull
the precomputed value for the suffix list to determine if, given the
decisions made so far, both beat the high score and satisfy the
configured caps. If we can't, we prune the entire subtree, reducing the
search space enormously.

This also adds a new variant of GetItemReforgeOptions that attempts to
only return reforge options that could possibly be useful, instead of
all the reforge options allowed by the game.

This has been tested on a survival hunter with 2 caps, a resto shaman
with 0 and 1 cap, an elemental shaman with 1 cap, and a mage with 1 cap.
The hunter shows a 20x speedup (60s -> 3s), the shamans complete almost
instantly no matter which solver is used, and the mage I am struggling
to figure out the performance regression for (it runs in 9s compared to
the dynamic programming solver's near-instantaneous solution).
Worst-case, this is just run after the DP takes too long or only when
two caps are configured, though for maintainability in the long-term
it'd be good to figure out one algorithm to rule them all.

There are two additional UI checkboxes - one to enable branch and bound,
and the other to do a spammy comparison of the two algorithms with a
bunch of debugging information to help diagnose when they diverge.